requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.3
importlib_metadata==8.5.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.6.0
scipy==1.13.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2024.2
Werkzeug==3.1.3
zipp==3.21.0
